# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.templateRequest:
    properties:
      is_aktif:
        example: 1
        type: integer
      isi:
        example: Halo {nama_customer}, pembayaran ke-{ke} sebesar Rp {nominal} telah
          kami terima.
        type: string
      nama:
        example: Kwitansi Pembayaran
        type: string
      tipe:
        example: kwitansi
        type: string
    required:
    - isi
    - nama
    - tipe
    type: object
...
  handlers.waRequest:
    properties:
      api_key:
        example: secret-key
        type: string
      api_url:
        example: https://wa.api.example.com
        type: string
      id_device:
        example: device-001
        type: string
      no_telp:
        example: "628123456789"
        type: string
    required:
    - api_url
    - id_device
    - no_telp
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@kavling-mentaya.com
    name: API Support
  description: REST API backend untuk sistem manajemen penjualan kavling tanah. Semua
    endpoint (kecuali /auth/login) memerlukan Bearer Token JWT di header Authorization.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Kavling Mentaya API
  version: "1.0"
paths:
...
  /konfigurasi:
    get:
      description: Mengambil data konfigurasi umum sistem (nama perusahaan, alamat,
        logo, tanda tangan, dll)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Ambil konfigurasi sistem
      tags:
      - Konfigurasi
    put:
      consumes:
      - multipart/form-data
      description: Mengubah data konfigurasi umum. Mendukung upload logo dan tanda
        tangan digital (multipart/form-data).
      parameters:
      - description: Nama perusahaan
        in: formData
        name: nama_perusahaan
        type: string
      - description: Alamat
        in: formData
        name: alamat
        type: string
      - description: Telepon
        in: formData
        name: telepon
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Website
        in: formData
        name: website
        type: string
      - description: Nama penanda tangan
        in: formData
        name: nama_ttd
        type: string
      - description: Jabatan penanda tangan
        in: formData
        name: jabatan_ttd
        type: string
      - description: File logo perusahaan
        in: formData
        name: logo
        type: file
      - description: File tanda tangan digital
        in: formData
        name: ttd_digital
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update konfigurasi sistem
      tags:
      - Konfigurasi
  /konfigurasi/template:
    get:
      description: Mengambil daftar semua template pesan WhatsApp
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List template pesan WhatsApp
      tags:
      - Konfigurasi
    post:
      consumes:
      - application/json
      description: Membuat template pesan WhatsApp baru. Variabel template menggunakan
        format {nama_variabel}.
      parameters:
      - description: Data template
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.templateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tambah template pesan WhatsApp
      tags:
      - Konfigurasi
  /konfigurasi/template/{id}:
    put:
      consumes:
      - application/json
      description: Mengubah template pesan WhatsApp berdasarkan ID
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data template
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.templateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update template pesan WhatsApp
      tags:
      - Konfigurasi
  /konfigurasi/wa:
    get:
      description: Mengambil konfigurasi gateway WhatsApp yang aktif
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Ambil konfigurasi WhatsApp
      tags:
      - Konfigurasi
    put:
      consumes:
      - application/json
      description: Mengubah atau membuat konfigurasi gateway WhatsApp
      parameters:
      - description: Konfigurasi WhatsApp
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.waRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update konfigurasi WhatsApp
      tags:
      - Konfigurasi
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}. Token diperoleh
      dari endpoint POST /api/auth/login.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"